Trait genmesh::generators::SharedVertex

pub trait SharedVertex<V>: Sized {
    fn shared_vertex(&self, i: usize) -> V;

    fn shared_vertex_count(&self) -> usize;

    fn shared_vertex_iter<'a>(&'a self) -> SharedVertexIterator<'a, Self, V> { ... }
}

The SharedVertex trait is meant to be used with the IndexedPolygon trait. This trait is meant as a way to calculate the shared vertices that are required to build the implementors mesh.

Required methods

fn shared_vertex(&self, i: usize) -> V

return the shared vertex at offset i

fn shared_vertex_count(&self) -> usize

return the number of shared vertices required to represent the mesh

Loading content...
Provided methods

fn shared_vertex_iter<'a>(&'a self) -> SharedVertexIterator<'a, Self, V>

create an iterator that returns each shared vertex that is required to build the mesh.

Loading content...