Trait genmesh::EmitLines

pub trait EmitLines {
    type Vertex;
    fn emit_lines<E>(self, emit: E)
    where
        E: FnMut(Line<Self::Vertex>);
}

Convert a Polygon into it's fragments

Associated Types

type Vertex

The Vertex defines the corners of a Polygon

Loading content...
Required methods

fn emit_lines<E>(self, emit: E) where
E: FnMut(Line<Self::Vertex>),

convert a polygon into lines, each line is emitted via calling of the callback of emit This allow for a variable amount of lines to be returned

Loading content...