Trait genmesh::EmitTriangles

pub trait EmitTriangles {
    type Vertex;
    fn emit_triangles<F>(&self, _: F)
    where
        F: FnMut(Triangle<Self::Vertex>);
}

provides a way to convert a polygon down to triangles

Associated Types

type Vertex

The content of each point in the face

Loading content...
Required methods

fn emit_triangles<F>(&self, _: F) where
F: FnMut(Triangle<Self::Vertex>),

convert a polygon to one or more triangles, each triangle is returned by calling emit

Loading content...