Trait genmesh::MapVertex

pub trait MapVertex<T, U> {
    type Output;
    fn map_vertex<F>(self, _: F) -> Self::Output
    where
        F: FnMut(T) -> U;
}

equivalent of map but per-vertex

Associated Types

type Output

Output should be a a container of the same shape of the type It's internal values should reflect any transformation the map did.

Loading content...
Required methods

fn map_vertex<F>(self, _: F) -> Self::Output where
F: FnMut(T) -> U,

map a function to each vertex in polygon creating a new polygon

Loading content...