Module genmesh::generators

A collection of utilties that can be used to build meshes programmatically.

Structs

Circle
Represents a circle in the XY plane with radius of 1, centered at (0, 0, 0)

Cone
The Cone mesh will create a mesh that is from 1 to -1 The bottom will be a circle around [0, 0, -1] with a radius of 1, all coords on the bottom will follow the plan equation -z-1=0 The tip of the cone will always be at coord [0, 0, 1]

Cube
A perfect cube, centered at (0, 0, 0) with each face starting at 1/-1 away from the origin

Cylinder
Represents a cylinder with radius of 1, height of 2, and centered at (0, 0, 0) pointing up (to 0, 0, 1).

IcoSphere
Icosahedral sphere with radius 1, centered at (0., 0., 0.)

IndexedPolygonIterator
An iterator that yields the indices of the mesh

Plane
Represents a 2D plane with origin of (0, 0), from 1 to -1

SharedVertexIterator
An iterator that yields the shared vertices of the mesh

SphereUv
Represents a sphere with radius of 1, centered at (0, 0, 0)

Torus

Traits

IndexedPolygon
The IndexedPolygon trait is used with the SharedVertex trait in order to build a mesh. IndexedPolygon calculates each polygon face required to build an implementors mesh. each face is always returned in indexed form that points to the correct vertice supplied by the SharedVertex trait.

SharedVertex
The SharedVertex trait is meant to be used with the IndexedPolygon trait. This trait is meant as a way to calculate the shared vertices that are required to build the implementors mesh.